Another procedure the example needs (~wlr_render_texture_with_matrix~).
Need to pull in the ~(bytestructure-class)~ module for
~bytestructure->pointer~ and ~bytestructure~ to handle the
arrays/vectors.

The original C code uses ~const float matrix[static 9]~ so, like the
compiler, we throw an error if the given list is too long (the error
message is similar to what =gcc= would throw but we use "list", in the
error message, rather than "array").