wlroots/types/matrix.scm: create file and add ~wlr-matrix-project-box~ procedure
The original C procedure expects the user to create an undefined array
variable to pass to the procedure for the procedure to fill with a
value.

Eschewing this (Guile/Scheme) unidiomatic behavior,
~wlr-matrix-project-box~ has the user give all the other variables,
instead, and returns a list.

The variable-to-set is defined as ~float mat[static 9]~ (and the
~projection~ array similarly) but the procedure only ever returns 9
values (and uses 9, from ~projection~).

So we hard-code the returned list to a length of 9 (the user's not
allowed to pass in a list, for ~projection~, that's shorter than 9, by
the compiler).

Like with ~wlr_render_texture_with_matrix~, we throw an error if
~projection~ is of a shorter length than 9.