wlroots/types/drm.scm: create file and add ~wlr-drm-create~
Made these when I was trying to create the =output-layers.c= example
before realizing that example only exists on the =master=
branch (right now) and, therefore, was using procedures and structs
which the version of =wlroots= that this project is using doesn't
have, yet.

No reason, I figure, to let work that's been done go to waste so
throwing this in here, as well.